requirements.txt
streamlit
google-generativeai
python-dotenv
psycopg2pip install -r equirements.txt